refactor: rewrite TouchPointer.java
